dvc-iris-ml-pipeline/requirements.txt
pandas
numpy
scikit-learn
pyyaml
dvc
